pairs work again
