+++ python/lsst/sims/featureScheduler/surveys/__init__.py
@@ -1,3 +1,4 @@
 from .base_survey import *
 from .surveys import *
 from .dd_surveys import *
+from .scripted_surveys import *
